+++ freppledb/input/commands/load.py
@@ -2145,7 +2145,8 @@ class loadOperationPlans(LoadTask):
                         case when operationplan.plan ? 'setupoverride'
                           then (operationplan.plan->>'setupoverride')::integer
                         end,
-                        coalesce(dmd.name, null)
+                        coalesce(dmd.name, null),
+                        remark
                         %s
                         FROM operationplan
                         LEFT OUTER JOIN (select name from demand
@@ -2167,8 +2168,8 @@ class loadOperationPlans(LoadTask):
                             dmd = frepple.demand(name=i[17])
                         elif with_fcst and i[18] and i[19]:
                             dmd = frepple.demand_forecastbucket(
-                                forecast=frepple.demand_forecast(name=i[18]),
-                                start=i[19],
+                                forecast=frepple.demand_forecast(name=i[19]),
+                                start=i[20],
                             )
                         else:
                             dmd = None
@@ -2186,6 +2187,7 @@ class loadOperationPlans(LoadTask):
                                 batch=i[13],
                                 quantity_completed=i[14],
                                 resources=i[15],
+                                remark=i[18],
                             )
                             if opplan:
                                 if i[5] == "confirmed":
@@ -2215,6 +2217,7 @@ class loadOperationPlans(LoadTask):
                                 source=i[6],
                                 create=create_flag,
                                 batch=i[13],
+                                remark=i[18],
                             )
                             if opplan and i[5] == "confirmed":
                                 if not consume_capacity:
@@ -2234,6 +2237,7 @@ class loadOperationPlans(LoadTask):
                                 source=i[6],
                                 create=create_flag,
                                 batch=i[13],
+                                remark=i[18],
                             )
                             if opplan:
                                 if i[5] == "confirmed":
@@ -2260,6 +2264,7 @@ class loadOperationPlans(LoadTask):
                                 source=i[6],
                                 create=create_flag,
                                 batch=i[13],
+                                remark=i[18],
                             )
                             if opplan:
                                 if i[5] == "confirmed":
@@ -2276,7 +2281,7 @@ class loadOperationPlans(LoadTask):
                             continue
 
                         if opplan:
-                            idx = 20 if with_fcst else 18
+                            idx = 21 if with_fcst else 19
                             for a in getAttributes(OperationPlan):
                                 setattr(opplan, a[0], i[idx])
                                 idx += 1
@@ -2303,7 +2308,7 @@ class loadOperationPlans(LoadTask):
                             where operationplan_id = operationplan.reference
                             order by resource_id
                         ),
-                        coalesce(dmd.name, null), coalesce(forecast.name, null), operationplan.due %s
+                        coalesce(dmd.name, null), remark, coalesce(forecast.name, null), operationplan.due %s
                         FROM operationplan
                         INNER JOIN (select reference
                         from operationplan %s
@@ -2347,7 +2352,8 @@ class loadOperationPlans(LoadTask):
                             where operationplan_id = operationplan.reference
                             order by resource_id
                         ),
-                        coalesce(dmd.name, null) %s
+                        coalesce(dmd.name, null),
+                        remark %s
                         FROM operationplan
                         INNER JOIN (select reference
                         from operationplan %s
@@ -2386,6 +2392,7 @@ class loadOperationPlans(LoadTask):
                             statusNoPropagation=i[5],
                             batch=i[8],
                             resources=i[9],
+                            remark=i[11],
                         )
                         if opplan:
                             if i[5] == "confirmed":
@@ -2406,12 +2413,12 @@ class loadOperationPlans(LoadTask):
                                     )
                             if i[10]:
                                 opplan.demand = frepple.demand(name=i[10])
-                            elif with_fcst and i[11] and i[12]:
+                            elif with_fcst and i[12] and i[13]:
                                 opplan.demand = frepple.forecastbucket(
-                                    forecast=frepple.demand_forecast(name=i[11]),
-                                    start=i[12],
+                                    forecast=frepple.demand_forecast(name=i[12]),
+                                    start=i[13],
                                 )
-                            idx = 13 if with_fcst else 11
+                            idx = 14 if with_fcst else 12
                             for a in getAttributes(OperationPlan):
                                 setattr(opplan, a[0], i[idx])
                                 idx += 1