+++ freppledb/input/migrations/0076_operationplan_remark.py
@@ -1,0 +1,53 @@
+#
+# Copyright (C) 2025 by frePPLe bv
+#
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+import datetime
+from django.db import migrations, models
+import django.db.models.deletion
+
+
+class Migration(migrations.Migration):
+    dependencies = [
+        ("input", "0075_operationmaterial_location"),
+    ]
+
+    operations = [
+        migrations.RunSQL(
+            "delete from common_attribute where name = 'remark' and model = 'operationplan'",
+            migrations.RunSQL.noop,
+        ),
+        migrations.RunSQL(
+            "alter table operationplan drop column if exists remark",
+            migrations.RunSQL.noop,
+        ),
+        migrations.AddField(
+            model_name="operationplan",
+            name="remark",
+            field=models.CharField(
+                blank=True,
+                help_text="remark",
+                max_length=300,
+                null=True,
+                verbose_name="remark",
+            ),
+        ),
+    ]