+++ src/tags.cpp
@@ -196,6 +196,7 @@ const Keyword Tags::quantity("quantity");
 const Keyword Tags::quantity_fixed("quantity_fixed");
 const Keyword Tags::quantity_completed("quantity_completed");
 const Keyword Tags::reference("reference");
+const Keyword Tags::remark("remark");
 const Keyword Tags::resource("resource");
 const Keyword Tags::resources("resources");
 const Keyword Tags::resourceskill("resourceskill");